05_02

+++ libft/ft_memset.c
@@ -1,7 +1,6 @@
-//#include <stdio.h>
 #include "libft.h"
 
-void *ft_memset(void *str, int c, /*size_t*/ unsigned int n)
+void *ft_memset(void *str, int c, size_t n)
 {
     int     i;
     char    ch_c;
